Move floater elasto aspect to elasto module
